db_connection_test/requirements.txt
jaydebeapi>=1.2.3
psycopg2-binary>=2.9.0
configparser 